+++ siriuspy/siriuspy/ramp/conn.py
@@ -16,7 +16,7 @@ from ..epics import EpicsProperty as _EpicsProperty, \
 from ..namesys import SiriusPVName as _PVName
 
 from ..pwrsupply.csdev import Const as _PSConst
-from ..csdevice.timesys import Const as _TIConst, \
+from ..timesys.csdev import Const as _TIConst, \
     get_hl_trigger_database as _get_trig_db
 from ..csdevice.orbitcorr import SOFBRings as _SOFBRings
 from ..search import LLTimeSearch as _LLTimeSearch, \
